requirements.txt
# Monitoring et métriques
influxdb-client>=1.36.1
psutil>=5.9.5
pandas>=2.2.0
numpy>=2.2.0

# Sécurité
cryptography>=41.0.3
python-dotenv>=1.0.0
pyotp>=2.8.0
qrcode>=7.4.2

# API et requêtes
requests>=2.31.0
aiohttp>=3.8.5
python-telegram-bot>=20.3

# Tests et validation
pytest>=7.4.0
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0

# Logging et gestion des erreurs
structlog>=23.1.0
tenacity>=8.2.2

# Visualisation
matplotlib>=3.7.2
seaborn>=0.12.2

# Optimisation
scipy>=1.10.1
scikit-learn>=1.3.0

# Utilitaires
python-dateutil>=2.8.2
pytz>=2023.3

# Trading
python-binance>=1.0.17
ccxt>=4.2.86
# TA-Lib>=0.4.28  # À installer séparément
backtesting>=0.3.2
redis>=5.0.1 